mock twitter credentials
